01-mcp-study/bonus_auth/pyproject.toml
[project]
name = "bonus-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.1",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "mcp[cli]>=1.12.3",
    "python-dotenv>=1.1.1",
]
